Overhaul of interactions with ATS system, fixed hacky install of test files, changed SPHERAL_TEST_INSTALL_PREFIX to include tests directory, added tests for viz and restart files in SpheralController, created separate performance analysis python file

+++ tests/functional/Hydro/Noh/Noh-cylindrical-2d.py
@@ -44,7 +44,6 @@ import os, shutil, mpi, sys
 from math import *
 
 from SolidSpheral2d import *
-from SpheralUtilities import adiak_value
 from SpheralTestUtilities import *
 from GenerateNodeDistribution2d import *
 from CubicNodeGenerator import GenerateSquareNodeDistribution
@@ -220,18 +219,21 @@ if asph:
 if solid:
     hydroname = "Solid" + hydroname
 
-dataDir = os.path.join(dataDir,
-                       hydroname,
-                       "nPerh=%f" % nPerh,
-                       "compatibleEnergy=%s" % compatibleEnergy,
-                       "Cullen=%s" % boolCullenViscosity,
-                       "filter=%f" % filter,
-                       "%s" % nodeMotion,
-                       "nrad=%i_ntheta=%i" % (nRadial, nTheta))
-restartDir = os.path.join(dataDir, "restarts")
-restartBaseName = os.path.join(restartDir, "Noh-cylindrical-2d-%ix%i" % (nRadial, nTheta))
-
-vizDir = os.path.join(dataDir, "visit")
+if dataDir:
+    dataDir = os.path.join(dataDir,
+                           hydroname,
+                           "nPerh=%f" % nPerh,
+                           "compatibleEnergy=%s" % compatibleEnergy,
+                           "Cullen=%s" % boolCullenViscosity,
+                           "filter=%f" % filter,
+                           "%s" % nodeMotion,
+                           "nrad=%i_ntheta=%i" % (nRadial, nTheta))
+    restartDir = os.path.join(dataDir, "restarts")
+    restartBaseName = os.path.join(restartDir, "Noh-cylindrical-2d-%ix%i" % (nRadial, nTheta))
+    vizDir = os.path.join(dataDir, "visit")
+else:
+    restartBaseName = None
+    vizDir = None
 if vizTime is None and vizCycle is None:
     vizBaseName = None
 else:
@@ -240,7 +242,7 @@ else:
 #-------------------------------------------------------------------------------
 # Check if the necessary output directories exist.  If not, create them.
 #-------------------------------------------------------------------------------
-if mpi.rank == 0:
+if mpi.rank == 0 and dataDir:
     if clearDirectories and os.path.exists(dataDir):
         shutil.rmtree(dataDir)
     if not os.path.exists(restartDir):
@@ -331,7 +333,6 @@ output("db")
 output("db.appendNodeList(nodes1)")
 output("db.numNodeLists")
 output("db.numFluidNodeLists")
-adiak_value("total_points", db.globalNumInternalNodes)
 
 #-------------------------------------------------------------------------------
 # Construct the hydro physics object.
