+++ tests/functional/Hydro/Noh/Noh-spherical-3d.py
@@ -7,7 +7,6 @@
 import os, shutil, sys
 from math import *
 from SolidSpheral3d import *
-from SpheralUtilities import adiak_value
 from SpheralTestUtilities import *
 from GenerateNodeDistribution3d import *
 
@@ -159,17 +158,20 @@ if asph:
 if solid:
     hydroname = "Solid" + hydroname
 
-dataDir = os.path.join(dataDir,
-                       hydroname,
-                       "nPerh=%f" % nPerh,
-                       "compatibleEnergy=%s" % compatibleEnergy,
-                       "Cullen=%s" % boolCullenViscosity,
-                       "filter=%f" % filter,
-                       "nx=%i_ny=%i_nz=%i" % (nx, ny, nz))
-restartDir = os.path.join(dataDir, "restarts")
-restartBaseName = os.path.join(restartDir, "Noh-spherical-3d-%ix%ix%i" % (nx, ny, nz))
-
-vizDir = os.path.join(dataDir, "visit")
+if dataDir:
+    dataDir = os.path.join(dataDir,
+                           hydroname,
+                           "nPerh=%f" % nPerh,
+                           "compatibleEnergy=%s" % compatibleEnergy,
+                           "Cullen=%s" % boolCullenViscosity,
+                           "filter=%f" % filter,
+                           "nx=%i_ny=%i_nz=%i" % (nx, ny, nz))
+    restartDir = os.path.join(dataDir, "restarts")
+    restartBaseName = os.path.join(restartDir, "Noh-spherical-3d-%ix%ix%i" % (nx, ny, nz))
+    vizDir = os.path.join(dataDir, "visit")
+else:
+    restartBaseName = None
+    vizDir = None
 if vizTime is None and vizCycle is None:
     vizBaseName = None
 else:
@@ -178,7 +180,7 @@ else:
 #-------------------------------------------------------------------------------
 # Check if the necessary output directories exist.  If not, create them.
 #-------------------------------------------------------------------------------
-if mpi.rank == 0:
+if mpi.rank == 0 and dataDir:
     if clearDirectories and os.path.exists(dataDir):
         shutil.rmtree(dataDir)
     if not os.path.exists(restartDir):
@@ -255,7 +257,6 @@ output("db")
 output("db.appendNodeList(nodes1)")
 output("db.numNodeLists")
 output("db.numFluidNodeLists")
-adiak_value("total_points", db.globalNumInternalNodes)
 
 #-------------------------------------------------------------------------------
 # Construct the hydro physics object.
